Error stability & Checker immutability  (#40)
* checker changed to const
* mnemonics map updated
* implemented variable substitution in macro params to be sublists
* removing warnings
* fixed an error in instruction map; allowed END ,
* removed variable.h warnings, grammar error with macro strings
* removed warnings parser_error_listener
* server handles files with uri 'untitled:Untitled-1'
* proc_grps.json can not be opened, no diagnostic is issued
* fixed error with rewinding to location we are currently at
* fixed workspace tests
* grammar std::transform issue fix
* added setting to disable continuation handling
* fixes
* using for std::function add_diag
* Update parser_library/src/processing/instruction_sets/low_language_processor.cpp
* changes in macro processor
* unassigned bool in lsp_context definition
* bad mnemonic transformation fix
* changed continuation handling to be dafaultly turned off
* added diagnostic_collector
* fixed string_to_macrodata method

+++ parser_library/src/context/location_counter.cpp
@@ -72,7 +72,7 @@ void location_counter::finish_layout(size_t offset)
 	assert(!(kind == loctr_kind::NONSTARTING) || (*spaces_[0]->name)[0] == 'B');
 
 	if (kind == loctr_kind::NONSTARTING)
-		space::resolve(spaces_[0], offset);
+		space::resolve(spaces_[0], (int)offset);
 
 	layuot_created_ = true;
 }
